requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.5
django-cors-headers==4.6.0
django-filter==25.1
django-sslserver==0.22
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
geographiclib==2.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
joblib==1.4.2
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
s3transfer==0.11.2
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
urllib3==2.3.0
